+++ dart/geodata/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 NOTE: [geodata version 1.4.0](https://github.com/navibyte/geospatial/milestone/7) currently under development (1.4.0-dev.0).
 
+🛠 Maintenance:
+* [Clarify ServiceException docs and add equality / hashCode methods #255](https://github.com/navibyte/geospatial/issues/255)
+
 ## 1.3.0
 
 ✨ New (2024-11-10): The stable version 1.3.0 with some refactoring (the